requirements.txt
osu.py==0.0.21
PyQt5~=5.15.2
humanize~=3.2.0